refactor (codemirror-dart) split editor.dart into multiple files
